Victoria Transport Data Platform/requirements.txt
# Data Engineering Dependencies
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0

# Data Processing
pyarrow>=10.0.0
pyspark>=3.4.0

# API Clients
google-transit-realtime-bindings>=1.0.0
protobuf>=4.0.0

# Data Visualization & Analysis
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Machine Learning (Optional)
scikit-learn>=1.2.0
prophet>=1.1.0

# Configuration & Logging
pyyaml>=6.0
python-dateutil>=2.8.0
pytz>=2022.7

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development Tools
black>=22.0.0
flake8>=5.0.0
jupyter>=1.0.0

# Schedule & Orchestration
schedule>=1.2.0
apache-airflow>=2.5.0  # Optional - for advanced orchestration
